[android] Added displaying traffic and compass buttons in PP and Search mode

+++ android/src/com/mapswithme/maps/NavigationButtonsAnimationController.java
@@ -54,7 +54,7 @@ class NavigationButtonsAnimationController
     });
   }
 
-  void setTopLimit(float limit)
+  void setTopLimit(int limit)
   {
     mTop = limit + mMargin;
   }
